refactor: separate mixins with an empty string, remove the gap in the footer and in the input on the timeline page, moved the url to env
