+++ src/components/MainPagesContent/styled.ts
@@ -4,5 +4,6 @@ import styled from 'styled-components';
 export const FallbackWrapper = styled.div`
   height: 100%;
   margin-top: 50px;
+
   ${mixinFlex({ alignItem: 'center', justifyContent: 'center' })};
 `;
